+++ src/steps/resource-manager/authorization/constants.ts
@@ -28,6 +28,10 @@ import {
   EOL_MATCHER,
   RESOURCE_GROUP_MATCHER,
 } from '../utils/matchers';
+import {
+  SUBSCRIPTION_ENTITY_METADATA,
+  STEP_RM_SUBSCRIPTIONS,
+} from '../subscriptions';
 
 // Fetch Role Assignments
 export const STEP_RM_AUTHORIZATION_ROLE_ASSIGNMENTS =
@@ -154,8 +158,8 @@ export const SCOPE_TYPES_MAP: ResourceIdMap[] = [
   ...RESOURCE_ID_TYPES_MAP,
   {
     resourceIdMatcher: new RegExp(SUBSCRIPTION_MATCHER + EOL_MATCHER),
-    _type: 'azure_subscription',
-    dependsOn: [],
+    _type: SUBSCRIPTION_ENTITY_METADATA._type,
+    dependsOn: [STEP_RM_SUBSCRIPTIONS],
   },
   {
     resourceIdMatcher: new RegExp(RESOURCE_GROUP_MATCHER + EOL_MATCHER),