+++ src/steps/resource-manager/resources/index.ts
@@ -2,6 +2,12 @@ import {
   Entity,
   Step,
   IntegrationStepExecutionContext,
+  ExplicitRelationship,
+  createDirectRelationship,
+  RelationshipClass,
+  IntegrationError,
+  StepRelationshipMetadata,
+  generateRelationshipType,
 } from '@jupiterone/integration-sdk-core';
 
 import { createAzureWebLinker } from '../../../azure';
@@ -13,7 +19,55 @@ import {
   RESOURCE_GROUP_ENTITY,
 } from './constants';
 import { createResourceGroupEntity } from './converters';
+import { SUBSCRIPTION_MATCHER } from '../utils/matchers';
+import {
+  SUBSCRIPTION_ENTITY_METADATA,
+  STEP_RM_SUBSCRIPTIONS,
+} from '../subscriptions';
 export * from './constants';
+
+const subscriptionRegex = new RegExp(SUBSCRIPTION_MATCHER);
+
+const SUBSCRIPTION_RESOURCE_GROUP_RELATIONSHIP_CLASS = RelationshipClass.HAS;
+const SUBSCRIPTION_RESOURCE_GROUP_RELATIONSHIP_METADATA: StepRelationshipMetadata = {
+  _class: SUBSCRIPTION_RESOURCE_GROUP_RELATIONSHIP_CLASS,
+  sourceType: SUBSCRIPTION_ENTITY_METADATA._type,
+  _type: generateRelationshipType(
+    SUBSCRIPTION_RESOURCE_GROUP_RELATIONSHIP_CLASS,
+    SUBSCRIPTION_ENTITY_METADATA._type,
+    RESOURCE_GROUP_ENTITY._type,
+  ),
+  targetType: RESOURCE_GROUP_ENTITY._type,
+};
+
+export async function createSubscriptionResourceGroupRelationship(
+  executionContext: IntegrationStepContext,
+  resourceGroupEntity: Entity,
+): Promise<ExplicitRelationship> {
+  const subscriptionIdMatch = resourceGroupEntity._key.match(subscriptionRegex);
+  if (!subscriptionIdMatch) {
+    throw new IntegrationError({
+      message: `Could not identify a subscription ID in the resource group _key: ${resourceGroupEntity._key}`,
+      code: 'UNMATCHED_SUBSCRIPTION',
+    });
+  }
+  const { jobState } = executionContext;
+  const subscriptionId = subscriptionIdMatch[0];
+
+  const subscriptionEntity = await jobState.findEntity(subscriptionId);
+  if (subscriptionEntity) {
+    return createDirectRelationship({
+      _class: RelationshipClass.HAS,
+      from: subscriptionEntity,
+      to: resourceGroupEntity,
+    });
+  } else {
+    throw new IntegrationError({
+      message: `Could not find the subscription "${subscriptionId}" in this integration.`,
+      code: 'MISSING_SUBSCRIPTION',
+    });
+  }
+}
 
 export async function fetchResourceGroups(
   executionContext: IntegrationStepContext,
@@ -30,6 +84,12 @@ export async function fetchResourceGroups(
       resourceGroup,
     );
     await jobState.addEntity(resourceGroupEntity);
+    await jobState.addRelationship(
+      await createSubscriptionResourceGroupRelationship(
+        executionContext,
+        resourceGroupEntity,
+      ),
+    );
   });
 }
 
@@ -40,8 +100,8 @@ export const resourcesSteps: Step<
     id: STEP_RM_RESOURCES_RESOURCE_GROUPS,
     name: 'Resource Groups',
     entities: [RESOURCE_GROUP_ENTITY],
-    relationships: [],
-    dependsOn: [STEP_AD_ACCOUNT],
+    relationships: [SUBSCRIPTION_RESOURCE_GROUP_RELATIONSHIP_METADATA],
+    dependsOn: [STEP_AD_ACCOUNT, STEP_RM_SUBSCRIPTIONS],
     executionHandler: fetchResourceGroups,
   },
 ];