+++ CHANGELOG.md
@@ -15,6 +15,8 @@ and this project adheres to
 - Added `azure_role_assignment|allows|<scope>` relationships
 - Added `azure_resource_group` entities
 - Added `azure_resource_group|has|<resource>` relationships
+- Added `azure_subscription` entities
+- Added `azure_subscription|has|azure_resource_group` relationships
 
 ### Changed